+++ src/channels/fixtures.ts
@@ -1,9 +1,9 @@
 import {
   AllocationStrategyEnum,
-  ChannelDetailsFragment,
   MarkAsPaidStrategyEnum,
   TransactionFlowStrategyEnum,
 } from "@dashboard/graphql";
+import { ChannelDetailsFragment } from "@dashboard/graphql/staging";
 
 export const channelsList: ChannelDetailsFragment[] = [
   {
@@ -47,6 +47,9 @@ export const channelsList: ChannelDetailsFragment[] = [
     },
     checkoutSettings: {
       automaticallyCompleteFullyPaidCheckouts: true,
+      allowLegacyGiftCardUse: true,
+      automaticCompletionDelay: 30,
+      automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
       __typename: "CheckoutSettings",
     },
   },
@@ -91,6 +94,9 @@ export const channelsList: ChannelDetailsFragment[] = [
     },
     checkoutSettings: {
       automaticallyCompleteFullyPaidCheckouts: true,
+      automaticCompletionDelay: 30,
+      automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
+      allowLegacyGiftCardUse: true,
       __typename: "CheckoutSettings",
     },
   },
@@ -135,6 +141,9 @@ export const channelsList: ChannelDetailsFragment[] = [
     },
     checkoutSettings: {
       automaticallyCompleteFullyPaidCheckouts: true,
+      allowLegacyGiftCardUse: true,
+      automaticCompletionDelay: 30,
+      automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
       __typename: "CheckoutSettings",
     },
   },
@@ -179,6 +188,9 @@ export const channelsList: ChannelDetailsFragment[] = [
     },
     checkoutSettings: {
       automaticallyCompleteFullyPaidCheckouts: true,
+      automaticCompletionDelay: 30,
+      automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
+      allowLegacyGiftCardUse: true,
       __typename: "CheckoutSettings",
     },
   },
@@ -223,6 +235,9 @@ export const channelsList: ChannelDetailsFragment[] = [
     },
     checkoutSettings: {
       automaticallyCompleteFullyPaidCheckouts: true,
+      allowLegacyGiftCardUse: true,
+      automaticCompletionDelay: 30,
+      automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
       __typename: "CheckoutSettings",
     },
   },
@@ -267,6 +282,9 @@ export const channelsList: ChannelDetailsFragment[] = [
     },
     checkoutSettings: {
       automaticallyCompleteFullyPaidCheckouts: true,
+      automaticCompletionDelay: 30,
+      automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
+      allowLegacyGiftCardUse: true,
       __typename: "CheckoutSettings",
     },
   },
@@ -311,6 +329,9 @@ export const channelsList: ChannelDetailsFragment[] = [
     },
     checkoutSettings: {
       automaticallyCompleteFullyPaidCheckouts: true,
+      automaticCompletionDelay: 30,
+      automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
+      allowLegacyGiftCardUse: true,
       __typename: "CheckoutSettings",
     },
   },
@@ -357,6 +378,9 @@ export const channel: ChannelDetailsFragment = {
   },
   checkoutSettings: {
     automaticallyCompleteFullyPaidCheckouts: true,
+    automaticCompletionDelay: 30,
+    automaticCompletionCutOffDate: "2024-01-01T00:00:00Z",
+    allowLegacyGiftCardUse: true,
     __typename: "CheckoutSettings",
   },
 };