+++ src/components/Date/index.ts
@@ -1,5 +1,0 @@
-export { default } from "./Date";
-export * from "./Date";
-export * from "./DateProvider";
-export * from "./DateTime";
-export * from "./FormatDate";