+++ src/giftCards/GiftCardCreateDialog/GiftCardCreateDialogContent.tsx
@@ -1,7 +1,7 @@
 // @ts-strict-ignore
 import { DashboardModal } from "@dashboard/components/Modal";
 import { GiftCardCreateInput, useGiftCardCreateMutation } from "@dashboard/graphql";
-import useCurrentDate from "@dashboard/hooks/useCurrentDate";
+import { useCurrentDate } from "@dashboard/hooks/useCurrentDate";
 import useNotifier from "@dashboard/hooks/useNotifier";
 import { DialogProps } from "@dashboard/types";
 import { useState } from "react";